labs/Lab1/requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
langchain-core>=0.1.0
python-dotenv>=1.0.0
